change and rename types
