+++ app/javascript/mastodon/features/emoji/utils.ts
@@ -1,6 +1,6 @@
 import EMOJI_REGEX from 'emojibase-regex/emoji-loose';
 
-export function stringHasEmoji(text: string): boolean {
+export function stringHasUnicodeEmoji(text: string): boolean {
   return EMOJI_REGEX.test(text);
 }
 
@@ -8,6 +8,6 @@ export function stringHasEmoji(text: string): boolean {
 const EMOJIS_FLAGS_REGEX =
   /[\u{1F1E6}-\u{1F1FF}|\u{E0062}-\u{E0063}|\u{E0065}|\u{E0067}|\u{E006C}|\u{E006E}|\u{E0073}-\u{E0074}|\u{E0077}|\u{E007F}]+/u;
 
-export function stringHasFlags(text: string): boolean {
+export function stringHasUnicodeFlags(text: string): boolean {
   return EMOJIS_FLAGS_REGEX.test(text);
 }
